refactor: move FakeMotor to TestModules and update PivotTest
