+++ src/test/java/frc/Glitch/Lib/BaseMechanisms/PivotTest.java
@@ -2,67 +2,13 @@ package frc.Glitch.Lib.BaseMechanisms;
 
 import Glitch.Lib.BaseMechanisms.Pivot;
 import Glitch.Lib.Motors.Motor;
+import frc.TestModules;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 public class PivotTest {
-    private static class FakeMotor implements Motor {
-        double positionRotations = 0.0; // rotations
-        double velocity = 0.0; // rotations/sec
-        double current = 0.0; // amps
-        double duty = 0.0; // -1..1
-        double lastFeedforward = 0.0;
-
-        @Override
-        public void setVelocity(double speed) {
-            this.velocity = speed;
-        }
-
-        @Override
-        public void setDutyCycle(double dutyCycle) {
-            this.duty = dutyCycle;
-        }
-
-        @Override
-        public void setPosition(double position, double feedforward) {
-            // For testing, snap immediately to requested position
-            this.positionRotations = position;
-            this.lastFeedforward = feedforward;
-        }
-
-        @Override
-        public void setPosition(double position) {
-            this.positionRotations = position;
-        }
-
-        @Override
-        public double getPosition() {
-            return positionRotations;
-        }
-
-        @Override
-        public double getCurrent() {
-            return current;
-        }
-
-        @Override
-        public double getVelocity() {
-            return velocity;
-        }
-
-        @Override
-        public boolean getForwardLimitSwitch() {
-            return false;
-        }
-
-        @Override
-        public boolean getReverseLimitSwitch() {
-            return false;
-        }
-    }
-
     private static class TestPivot extends Pivot {
         public TestPivot(Motor motor) {
             super(
@@ -75,12 +21,12 @@ public class PivotTest {
         }
     }
 
-    private FakeMotor motor;
+    private TestModules.FakeMotor motor;
     private TestPivot pivot;
 
     @BeforeEach
     void setUp() {
-        motor = new FakeMotor();
+        motor = new TestModules.FakeMotor();
         pivot = new TestPivot(motor);
     }
